Authorization with Gates and use debugbar package to follow performence

+++ store/app/Http/Controllers/Admin/CategoryController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use App\Models\Category;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Str;
 
@@ -15,6 +16,10 @@ class CategoryController extends Controller
      */
     public function index(Request $request)
     {
+        if (!Gate::allows('categories.view'))
+        {
+            abort(403);
+        }
         //SELECT child.* ,parent.name as parent_name FROM categories as child
         //leftjoin categories as parent ON  parent.id = child.parent_id
         $categories = Category::with('parent')
@@ -37,6 +42,10 @@ class CategoryController extends Controller
      */
     public function create()
     {
+        if (Gate::denies('categories.create'))
+        {
+            abort(403);
+        }
         $parents = Category::all();
         $category = new Category();
         return view("dashboard.categories.create",compact('parents','category'));
@@ -48,6 +57,10 @@ class CategoryController extends Controller
      */
     public function store(Request $request)
     {
+        if (Gate::denies('categories.create'))
+        {
+            abort(403);
+        }
         $request->validate(Category::rules());
         $data = $request->except('image');
         $request->merge(['slug'=>Str::slug($request->post('name'))]);
@@ -65,6 +78,7 @@ class CategoryController extends Controller
      */
     public function show(Category $category)
     {
+        Gate::authorize('categories.view');
         return view('dashboard.categories.show',compact('category'));
     }
 
@@ -73,6 +87,7 @@ class CategoryController extends Controller
      */
     public function edit(string $id)
     {
+        Gate::authorize('categories.update');
         //SELECT * FROM categories WHERE id <> $id AND
         //(parent_id IS NULL OR parent_id <> $id)
         $parents = Category::where('id','<>',$id)
@@ -90,6 +105,8 @@ class CategoryController extends Controller
      */
     public function update(Request $request, string $id)
     {
+        Gate::authorize('categories.update');
+
         $request->validate(Category::rules($id));
         $category = Category::findorFail($id);
         $old_image = $category->image;
@@ -114,6 +131,8 @@ class CategoryController extends Controller
      */
     public function destroy(Category $category)
     {
+        Gate::authorize('categories.delete');
+
         $category->delete();
         return redirect()->route('dashboard.categories.index')
             ->with('success','Category deleted!');
